Move enums into a separate module
